MultipartPost test now uses the form feature
The previous implementation relied on manually reading the body and using a MultipartReader object in order to fully consume all the parts and read all the bytes. This is actually redundant as ASP.NET has a built-in feature where multiparts are read and cached in a local file even before the controller is called. All the controller has to do is either specify a complex object as its parameter or access the files via the Request.Form object.