Add test KestrelShutdown
This is intended to uncover problems when starting and stopping Kestrel is done repeatedly. A user might choose to restart Kestrel within the same process, or restart the process entirely. This code is currently restarting Kestrel within the same process but it's easi to modify this and test restarting the process instead.